+++ src/day07/InvalidValueException.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * Created by PhpStorm.
+ *
+ * PHP version 5.6, 7
+ *
+ * @package Zeravcic\TddByJohnCleary\day07
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+
+
+namespace Zeravcic\TddByJohnCleary\day07;
+
+
+/**
+ * Class RecentlyUsedList
+ *
+ * @package Zeravcic\TddByJohnCleary\day07
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+class InvalidValueException extends \Exception
+{
+}